backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3
pdfplumber==0.11.4
python-docx==1.1.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
